pets: extract pets module
