+++ pets/src/test/java/net/kemitix/naolo/pets/AddPetTest.java
@@ -1,8 +1,6 @@
-package net.kemitix.naolo.core.pets;
+package net.kemitix.naolo.pets;
 
 import net.kemitix.naolo.entities.Owner;
-import net.kemitix.naolo.entities.Pet;
-import net.kemitix.naolo.entities.PetType;
 import net.kemitix.naolo.storage.EntityRepository;
 import org.assertj.core.api.WithAssertions;
 import org.junit.jupiter.api.DisplayName;
