+++ pets/src/test/java/net/kemitix/naolo/pets/PetRepositoryTest.java
@@ -1,6 +1,6 @@
-package net.kemitix.naolo.storage;
+package net.kemitix.naolo.pets;
 
-import net.kemitix.naolo.entities.Pet;
+import net.kemitix.naolo.storage.EntityRepository;
 import org.assertj.core.api.WithAssertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
