+++ pom.xml
@@ -10,6 +10,7 @@
         <module>testing</module>
         <module>core</module>
         <module>visits</module>
+        <module>pets</module>
         <module>runner-quarkus</module>
         <module>api-tests</module>
         <module>war</module>
