+++ visits/src/main/java/net/kemitix/naolo/plugin/visits/Visit.java
@@ -3,8 +3,8 @@ package net.kemitix.naolo.plugin.visits;
 
 import lombok.*;
 import net.kemitix.naolo.entities.JPAActivityListener;
-import net.kemitix.naolo.entities.Pet;
 import net.kemitix.naolo.entities.Veterinarian;
+import net.kemitix.naolo.pets.Pet;
 import net.kemitix.naolo.storage.HasId;
 
 import javax.json.bind.annotation.JsonbDateFormat;
