+++ visits/src/test/java/net/kemitix/naolo/plugin/visits/VisitTest.java
@@ -1,7 +1,7 @@
 package net.kemitix.naolo.plugin.visits;
 
-import net.kemitix.naolo.entities.Pet;
 import net.kemitix.naolo.entities.Veterinarian;
+import net.kemitix.naolo.pets.Pet;
 import org.assertj.core.api.WithAssertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
