+++ war/pom.xml
@@ -26,16 +26,6 @@
         </dependency>
         <dependency>
             <groupId>net.kemitix.naolo</groupId>
-            <artifactId>naolo-core</artifactId>
-            <exclusions>
-                <exclusion>
-                    <groupId>org.projectlombok</groupId>
-                    <artifactId>lombok</artifactId>
-                </exclusion>
-            </exclusions>
-        </dependency>
-        <dependency>
-            <groupId>net.kemitix.naolo</groupId>
             <artifactId>naolo-visits</artifactId>
             <exclusions>
                 <exclusion>
