+++ war/src/test/java/net/kemitix/naolo/war/VisitsIT.java
@@ -1,11 +1,11 @@
 package net.kemitix.naolo.war;
 
 import net.kemitix.naolo.OwnerResource;
-import net.kemitix.naolo.PetResource;
 import net.kemitix.naolo.VetResource;
 import net.kemitix.naolo.entities.Owner;
-import net.kemitix.naolo.entities.Pet;
 import net.kemitix.naolo.entities.Veterinarian;
+import net.kemitix.naolo.pets.Pet;
+import net.kemitix.naolo.pets.PetResource;
 import net.kemitix.naolo.plugin.visits.Visit;
 import net.kemitix.naolo.plugin.visits.VisitResource;
 import org.assertj.core.api.WithAssertions;
